move serializers for user data
